[FEATURE] Support multiple listeners by using advertisedListeners.
Suppose we deploy kop with pulsar on a host which have two NIC, such as a cloud ECS. 100.73.36.31:9092 is the address for private network inside the cloud and 100.71.36.31 is the address for public network. If we support multiple addresses for access to a protocol, we can do the following configuration.
Multiple Listeners configuration examples:

kafkaProtocolMap=EXTERNAL:PLAINTEXT,INTERNAL:PLAINTEXT
kafkaListeners=EXTERNAL://100.71.36.31:9092,INTERNAL://100.73.36.31:9093
kafkaAdvertisedListeners=PLAINTEXT://100.71.36.31:9092